読みモデル重みを CSC 形式で保存しロード時のピークメモリを削減
推論は特徴量(列)方向で走査するため CSC が最終形だが、ファイルは CSR
(行=クラス)で保存しており、ロード時の CSR→CSC 変換で両方が同時にメモリ上に
存在してピークが跳ねていた。ファイル自体を CSC にして変換を撤去する。
sklearn の coef_ が行=クラスのレイアウトで、量子化 scale もクラスごとに
行単位で計算するため CSR が自然だったが、転置のコストをロード時に先送り
していたことになる。量子化は行が連続する CSR で済ませ、転置だけを
エクスポート時に行うようにした。

行インデックスはクラスID (< n_classes) なので u16 で持てる。非ゼロ要素
1個あたり値(int8)と対で常駐するため、u32 のままでは行インデックスだけで
値の4倍を占め、int8 量子化の効果を打ち消していた。

実測 (window=4/5/7):
  ピークメモリ  24.9→20.2 / 44.3→36.3 / 64.4→53.1 MB
  ファイル      10.0→9.3  / 20.4→19.4 / 30.4→29.7 MB
点訳出力は変更前と完全一致(3モデル × .mbm/.mbmf で確認)。列ポインタが
n_classes+1 個から n_features+1 個に増えるため、ファイルサイズの利得は
nnz/n_features 比に依存する(2 が損益分岐)。

あわせて .mbmf のバージョン番号を .mbm と共有する採番に統一し
(独立採番の 0x02 → 0x05)、VERSION / MAX_CLASSES / MAX_REASONABLE_COUNT を
loader.rs 一箇所に集約した。複製すると次のフォーマット変更で片方だけ
更新してしまうため。2フォーマットの区別は magic のみで行う。

旧 v4 モデルは明示的にエラーになるので再エクスポートが必要。

+++ momors/crates/momors-core/src/float_model.rs
@@ -32,7 +32,7 @@ pub struct FloatMomoModel {
     /// CSC 列ポインタ (size: `n_features + 1`)
     pub(crate) csc_colptr: Vec<u32>,
     /// CSC 行インデックス = クラスID (size: `n_nonzero`)
-    pub(crate) csc_rowind: Vec<u32>,
+    pub(crate) csc_rowind: Vec<u16>,
     /// CSC 非ゼロ値 (size: `n_nonzero`)。量子化なしの実値そのもの。
     pub(crate) csc_data: Vec<f32>,
